split Log4netSyslogSeverityConvertor to a separate file
